+++ src/vue/App.vue
@@ -3,20 +3,6 @@
     id="app"
     v-if="isAppInitialized"
     :key="lang">
-    <warning-banner
-      v-if="isNotSupportedBrowser && !isSupportedBrowsersPage"
-      :message="'common.browser-not-supported' | globalize"
-      message-type="warning"
-    >
-      <router-link
-        tag="a"
-        :to="vueRoutes.supportedBrowsers.name"
-        class="app__warning-message-link"
-      >
-        {{ 'warning-banner.supported-browsers-list' | globalize }}
-      </router-link>
-    </warning-banner>
-
     <template v-if="isLoggedIn && isNavigationRendered">
       <warning-banner
         v-if="isAccountBlocked"
@@ -54,8 +40,6 @@ import Navbar from '@/vue/navigation/Navbar.vue'
 import Sidebar from '@/vue/navigation/Sidebar.vue'
 import WarningBanner from '@/vue/common/WarningBanner'
 
-import { isCompatibleBrowser } from '@/js/helpers/is-compatible-browser'
-
 import {
   mapGetters,
   mapActions,
@@ -84,7 +68,6 @@ export default {
   },
 
   data: () => ({
-    isNotSupportedBrowser: false,
     isAppInitialized: false,
     vueRoutes,
     lang: i18n.language,
@@ -102,9 +85,6 @@ export default {
     isNavigationRendered () {
       return this.$route.matched.some(m => m.meta.isNavigationRendered)
     },
-    isSupportedBrowsersPage () {
-      return this.$route.name === vueRoutes.supportedBrowsers.name
-    },
   },
 
   watch: {
@@ -121,7 +101,6 @@ export default {
 
     this.startIdle()
 
-    this.detectIncompatibleBrowser()
     i18n.onLanguageChanged(lng => (this.lang = lng))
 
     this.watchChangesInLocalStorage()
@@ -159,9 +138,6 @@ export default {
       factorsManager.useApi(api)
       documentsManager.useApi(api)
       await this.loadAssets()
-    },
-    detectIncompatibleBrowser () {
-      this.isNotSupportedBrowser = !isCompatibleBrowser()
     },
 
     async getDecryptedSecretSeed () {