+++ test.config.js
@@ -1,7 +1,8 @@
 require('jsdom-global')('', { url: 'http://localhost' })
 
 global.localStorage = window.localStorage
-global.expect = require('chai').expect
 global.sinon = require('sinon')
+
 const chai = require('chai')
-chai.use(require('sinon-chai'));
+global.expect = chai.expect
+chai.use(require('sinon-chai'))